requirements.txt
openai>=2.24.0
openai-agents>=0.10.1
python-dotenv>=1.0.1
pydantic>=2.8.2
pytest>=8.3.2
google-api-python-client>=2.169.0
google-auth>=2.40.0
google-auth-oauthlib>=1.2.0
redis>=5.0.0
gunicorn